feat: add v4 capabilities and adapt tests
To use v4 endpoints, pass in 'useV4' = true in LeagueJS constructor.
This change is backwards-compatible.
There will be a major release due to the breaking nature of the chages,
in which v4 will be used by default.

* add apiVersionOverrides to options/config
This can be used to config which version which endpoint is using.
API_VERSION will be used, if there is no override provided.
For now, the overrides are only used if `useV4` is set in the endpoint's config.
This is passed down from the LeagueJS constructor options object.

* Mock summoner moved to testConfig
Because the tests require encrypted values, for now the mock summoner needs to be
provided with encrypted values generated by the used API-KEY.
The encrypted values will stay the same if the API-key is regenerated, as long as the API-KEY is a
Dev-key from the same account or the same project it's on the same account or the same project

* temporary checking id parameters for string and number type, depending on apiVersion